+++ .eslintignore
@@ -2,3 +2,4 @@ node_modules
 build
 test
 coverage 
+serve
