+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "el-tag",
-  "version": "1.0.1",
+  "version": "1.0.0",
   "description": "Handy utility to create DOM elements using Tagged Template Literals and retrieving references.",
   "main": "index.js",
   "author": "Victor Cano",
@@ -10,6 +10,7 @@
     "build": "run-s clean && run-p build:*",
     "build:main": "tsc -p tsconfig.json",
     "build:module": "tsc -p tsconfig.module.json",
+    "build:serve": "tsc -p tsconfig.serve.json",
     "fix": "run-s fix:*",
     "fix:prettier": "prettier \"src/**/*.ts\" --write",
     "fix:eslint": "eslint --fix . --ext .js,.jsx,.ts,.tsx",
@@ -17,6 +18,7 @@
     "test:lint": "eslint . --ext .js,.jsx,.ts,.tsx",
     "test:unit": "nyc --silent ava",
     "watch": "run-s clean build:main && run-p \"build:main -- -w\" \"test:unit -- --watch\"",
+    "chrome": "run-s clean build:serve && run-p serve \"build:serve -- -w\" \"test:unit -- --watch\"",
     "cov": "run-s build test:unit cov:html && open-cli coverage/index.html",
     "cov:html": "nyc report --reporter=html",
     "cov:send": "nyc report --reporter=lcov && codecov",
@@ -28,7 +30,8 @@
     "doc:publish": "gh-pages -m \"[ci skip] Updates\" -d build/docs --dotfiles ",
     "version": "standard-version",
     "reset": "git clean -dfx && git reset --hard && yarn install",
-    "clean": "trash build test",
+    "clean": "trash build test serve/out",
+    "serve": "serve serve",
     "prepare-release": "run-s reset test cov:check doc:html version doc:touch doc:publish",
     "preinstall": "node -e \"if(process.env.npm_execpath.indexOf('yarn') === -1) throw new Error('el tag must be installed with Yarn: https://yarnpkg.com/')\""
   },
@@ -62,6 +65,7 @@
     "nyc": "^15.1.0",
     "open-cli": "^6.0.1",
     "prettier": "^2.0.5",
+    "serve": "^11.3.2",
     "standard-version": "^8.0.2",
     "touch-cli": "^0.0.1",
     "trash-cli": "^3.0.0",
@@ -71,17 +75,22 @@
   "engines": {
     "node": ">=14"
   },
+  "type": "module",
   "ava": {
     "verbose": true,
     "failFast": true,
     "files": [
-      "build/main/**/*.spec.js"
+      "build/main/**/*.spec.js",
+      "build/module/**/*.spec.js"
     ],
     "ignoredByWatcher": [
-      "build/main/**/*.js"
+      "build/main/**/*.js",
+      "build/module/**/*.spec.js"
     ],
-    "require": [
-      "./src/test/_setup-browser-env.js"
+    "require": [ 
+      "./src/test/_setup-browser-env.js",
+      "babel-register"
+      
     ]
   },
   "prettier": {
@@ -92,5 +101,8 @@
     "exclude": [
       "**/*.spec.js"
     ]
+  },
+  "dependencies": {
+    "babel-register": "^6.26.0"
   }
 }
