+++ serve/index.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Document</title>
+
+    <script>
+      var exports = {};
+    </script>
+
+    <script type="module" src="/index.js"></script>
+ 
+  </head>
+  <body></body>
+</html>
