+++ serve/index.js
@@ -1,0 +1,4 @@
+import elTag from './out/index.js';
+ 
+
+window.elTag = elTag;
