+++ tsconfig.serve.json
@@ -1,0 +1,11 @@
+{
+  "extends": "./tsconfig",
+  "compilerOptions": {
+    "target": "ES6",
+    "module": "ES6",
+    "outDir": "serve/out"
+  },
+  "exclude": [
+    "node_modules/**" 
+  ]
+}
